refactor(test): extract test utils
