+++ test/utils.js
@@ -1,0 +1,62 @@
+/*
+ * Copyright 2019 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+const crypto = require('crypto');
+const path = require('path');
+const fse = require('fs-extra');
+const util = require('util');
+
+async function createTestRoot() {
+  const dir = path.resolve(__dirname, 'tmp', crypto.randomBytes(16)
+    .toString('hex'));
+  await fse.ensureDir(dir);
+  return dir;
+}
+exports.createTestRoot = createTestRoot;
+
+const ANSI_REGEXP = RegExp([
+  '[\\u001B\\u009B][[\\]()#;?]*(?:(?:(?:[a-zA-Z\\d]*(?:;[a-zA-Z\\d]*)*)?\\u0007)',
+  '(?:(?:\\d{1,4}(?:;\\d{0,4})*)?[\\dA-PRZcf-ntqry=><~]))',
+].join('|'), 'g');
+
+class TestLogger {
+  constructor() {
+    this.messages = [];
+  }
+
+  _log(level, ...args) {
+    this.messages.push(util.format(...args).replace(ANSI_REGEXP, ''));
+    // eslint-disable-next-line no-console
+    console[level](...args);
+  }
+
+  get output() {
+    return this.messages.join('\n');
+  }
+
+  debug(...args) {
+    this._log('debug', ...args);
+  }
+
+  info(...args) {
+    this._log('info', ...args);
+  }
+
+  warn(...args) {
+    this._log('warn', ...args);
+  }
+
+  error(...args) {
+    this._log('error', ...args);
+  }
+}
+
+exports.TestLogger = TestLogger;
